Move recipe metaclases into its own module
