+++ numina/core/recipes.py
@@ -37,7 +37,7 @@ from .recipeinout import RecipeRequirements as RecipeRequirementsClass
 _logger = logging.getLogger('numina')
 
 
-class _BaseRecipeMethods(object):
+class BaseRecipeMethods(object):
     '''Base class for all instrument recipes'''
 
     RecipeResult = RecipeResultClass
@@ -47,7 +47,7 @@ class _BaseRecipeMethods(object):
     logger = _logger
 
     def __init__(self, *args, **kwds):
-        super(_BaseRecipeMethods, self).__init__()
+        super(BaseRecipeMethods, self).__init__()
         self.__author__ = 'Unknown'
         self.__version__ = '0.0.0'
         # These two are maintained
@@ -76,7 +76,7 @@ class _BaseRecipeMethods(object):
         return cls.RecipeResult(*args, **kwds)
 
     def run(self, recipe_input):
-        return self.RecipeResult()
+        return self.create_result()
 
     def __call__(self, recipe_input):
         '''
@@ -108,84 +108,11 @@ class _BaseRecipeMethods(object):
         return hdr
 
 
-class BaseRecipe(_BaseRecipeMethods):
+class BaseRecipe(BaseRecipeMethods):
     '''Base class for all instrument recipes'''
 
     __metaclass__ = abc.ABCMeta
 
     @abc.abstractmethod
     def run(self, recipe_input):
-        return self.RecipeResult()
-
-
-from .recipeinout import RecipeResult, RecipeRequirements
-from .recipeinout import RecipeResultAutoQC
-from .dataholders import Product
-from .requirements import Requirement
-
-
-class RecipeType(type):
-    '''Metaclass for Recipe.'''
-    def __new__(cls, classname, parents, attributes):
-        filter_reqs = {}
-        filter_prods = {}
-        filter_attr = {}
-
-        for name, val in attributes.items():
-            if isinstance(val, Requirement):
-                filter_reqs[name] = val
-            elif isinstance(val, Product):
-                filter_prods[name] = val
-            else:
-                filter_attr[name] = val
-
-        ReqsClass = cls.create_req_class(classname, filter_reqs)
-
-        ResultClass = cls.create_prod_class(classname, filter_prods)
-
-        filter_attr['RecipeResult'] = ResultClass
-        filter_attr['RecipeRequirements'] = ReqsClass
-        return super(RecipeType, cls).__new__(
-            cls, classname, parents, filter_attr)
-
-    @classmethod
-    def create_req_class(cls, classname, attributes):
-        if attributes:
-            reqs_name = '%sRequirements' % classname
-            ReqsClass = type(reqs_name, (RecipeRequirements,), attributes)
-        else:
-            ReqsClass = RecipeRequirements
-        return ReqsClass
-
-    @classmethod
-    def create_prod_class(cls, classname, attributes):
-        if attributes:
-            result_name = '%sResult' % classname
-            ResultClass = type(result_name, (RecipeResult,), attributes)
-        else:
-            ResultClass = RecipeResult
-
-        return ResultClass
-
-
-class RecipeTypeAutoQC(RecipeType):
-    '''Metaclass for Recipe with RecipeResultAutoQC.'''
-    @classmethod
-    def create_prod_class(cls, classname, attributes):
-        if attributes:
-            result_name = '%sResult' % classname
-            ResultClass = type(result_name, (RecipeResultAutoQC,), attributes)
-        else:
-            ResultClass = RecipeResult
-
-        return ResultClass
-
-
-class BaseRecipeAlt(_BaseRecipeMethods):
-    '''Base class for instrument recipes'''
-    __metaclass__ = RecipeType
-
-
-class BaseRecipeAutoQC(_BaseRecipeMethods):
-    '''Base class for instrument recipes'''
-    __metaclass__ = RecipeTypeAutoQC
+        return self.create_result()
